requirements.txt
discord.py
python-decouple
logging
fastapi
uvicorn
asyncio
redis
python-multipart
aiofiles